% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-noise-scale.R
\name{catoni_influence_fun}
\alias{catoni_influence_fun}
\title{Influence function from Catoni (2012)}
\usage{
catoni_influence_fun(xx)
}
\arguments{
\item{xx}{data vector}
}
\description{
Influence function from Catoni (2012)
}
\references{
Catoni, Olivier. "Challenging the empirical mean and empirical variance: a deviation study." Annales de l'IHP Probabilités et statistiques. Vol. 48. No. 4. 2012.

McGonigle, Euan T., and Haeran Cho. "Robust multiscale estimation of time-average variance for time series segmentation." Computational Statistics & Data Analysis 179 (2023): 107648.

https://github.com/EuanMcGonigle/TAVC.seg
}
