% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised-tacv-est.R
\name{catoni_sum}
\alias{catoni_sum}
\title{Funtion to be optimised to obtain an estimate for \tau}
\usage{
catoni_sum(tau, xx, qq)
}
\arguments{
\item{tau, }{parameter}

\item{xx, }{data vector}

\item{qq, }{scaling}
}
\description{
Funtion to be optimised to obtain an estimate for \tau
}
\references{
Catoni, Olivier. "Challenging the empirical mean and empirical variance: a deviation study." Annales de l'IHP Probabilités et statistiques. Vol. 48. No. 4. 2012.

McGonigle, Euan T., and Haeran Cho. "Robust multiscale estimation of time-average variance for time series segmentation." Computational Statistics & Data Analysis 179 (2023): 107648.

https://github.com/EuanMcGonigle/TAVC.seg
}
