% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_perturbed_data_tL_tR}
\alias{construct_perturbed_data_tL_tR}
\title{Construct perturbed data around an estimated changepoint}
\usage{
construct_perturbed_data_tL_tR(y, tL, thj, tR, phi)
}
\arguments{
\item{y}{observed data}

\item{tL}{left end point}

\item{thj}{estimated changepoint}

\item{tR}{right end point}

\item{phi}{perturb constant}
}
\value{
\code{yphi} perturbed data
}
\description{
Construct perturbed data around an estimated changepoint
}
\details{
Perturbs the observed data y around thj based on a scalar phi:

\deqn{y'(\phi) = y - v * t(v) * y / ||v||_2^2 + v * \phi / ||v||_2^2}

See \code{\link{construct_v_tL_tR}} for description of tL and tR.
}
\examples{

### Generate sample data
set.seed(1)
mu <- rep(c(1, 3, -4), each = 100)
dat <- mu + rnorm(length(mu))

# Suppose a changepoint is estimated at thj = 100. Then a fixed window contrast around
# thj with window size = 10 is constructed as follows:
thj <- 100
window_size <- 10
v <- construct_v_tL_tR(thj - window_size + 1, thj, thj + window_size, length(dat))

# Perturb the data around thj
phi <- 2
yphi <- dat - v \%*\% t(v) \%*\% dat / (sum(v * v)) + v * phi / sum(v * v)

# construct_perturbed_data_tL_tR is just a utility function for the above calculation:
sum(yphi - construct_perturbed_data_tL_tR(dat, thj - window_size + 1, thj, thj + window_size, phi))

}
\references{
Jewell, S., Fearnhead, P., and Witten, D. (2019+). Testing for a
change in mean after changepoint detection. Technical report.
}
\seealso{
\code{\link{changepoint_estimates}},
\code{\link{changepoint_inference}}, \code{\link{construct_v_tL_tR}}
}
