% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_v_tL_tR}
\alias{construct_v_tL_tR}
\title{Construct contrast around an estimated changepoint}
\usage{
construct_v_tL_tR(tL, thj, tR, n)
}
\arguments{
\item{tL}{left end point}

\item{thj}{estimated changepoint to construct contrast vector around}

\item{tR}{right end point}

\item{n}{number of observed datapoints}
}
\value{
\code{nu} contrast vector
}
\description{
Construct contrast around an estimated changepoint
}
\details{
The ith entry of the n-vector contrast is defined as:

0,                    i < tL or i > tR

1 / \eqn{(\hat{\tau}_j - tL + 1)},    \eqn{tL \le i \le \hat{\tau}_j}

-1 / \eqn{tR - \hat{\tau}_j},       \eqn{\hat{\tau}_j + 1 \le i \le tR}

where tL = max(1, \eqn{\hat{\tau}}_j - h + 1) and tR = min(T,
\eqn{\hat{\tau}_j} + h) in the fixed window case with window size = h, and
where tL = max(1, \eqn{\hat{\tau}_{j-1}} + 1) and tR = min(T,
\eqn{\hat{\tau}_{j + 1}}) in the adaptive case.
}
\examples{

### Generate sample data
set.seed(1)
mu <- rep(c(1, 3, -4), each = 100)
dat <- mu + rnorm(length(mu))

# Suppose a changepoint is estimated at thj = 100. Then a fixed window contrast around
# thj with window size = 10 is constructed as follows:
thj <- 100
window_size <- 10
v <- construct_v_tL_tR(thj - window_size + 1, thj, thj + window_size, length(dat))

# The change in mean using the underlying mean around thj is (1 - 3 = -2)
sum(v * mu)

# Using the observed data
sum(v * dat)

}
\references{
Jewell, S., Fearnhead, P., and Witten, D. (2019+). Testing for a change in mean after changepoint detection. Technical report.
}
\seealso{
\code{\link{changepoint_estimates}}, \code{\link{changepoint_inference}}
}
