% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt}
\alias{cpt}
\title{Change point locations}
\usage{
cpt(obj, cpt_loc = c("midpoint", "RSS")[1])
}
\arguments{
\item{obj}{An object of class 'cptInference', returned by \code{diffInf}.}
}
\description{
Estimate change point locations
}
