% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffInf.R
\name{diffInf}
\alias{diffInf}
\title{Change point inference via differencing}
\usage{
diffInf(
  xx,
  degree,
  alpha = 0.1,
  gaussian_noise = TRUE,
  independent_noise = TRUE,
  tau = NULL,
  aa = sqrt(2),
  min_scale = floor(sqrt(length(xx))),
  HH = NULL
)
}
\arguments{
\item{xx}{signal to be segmented}

\item{degree}{polynomial degree of the underlying signal}

\item{alpha}{desired coverage}

\item{gaussian_noise}{whether the noise is Gaussian}

\item{independent_noise}{whether the noise is serially dependent}

\item{tau}{noise level if know, if null the noise level will be estimated via MAD if the noise in independent and via TACV estimator if the noise is dependent}

\item{aa}{controls grid density (a in the paper)}

\item{min_scale}{minimum width of intervals tested for a change}

\item{HH}{supply numeric constant H if pre-compute}
}
\description{
TO-DO!
}
