% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_paths.R
\name{estimate_paths}
\alias{estimate_paths}
\title{Estimate changepoints using
automatic tuning parameter selection within
a range of values [lambda_min, lambda_max]}
\usage{
estimate_paths(dat, lambda_min = 0.01, lambda_max = 10,
  max_iters = 10)
}
\arguments{
\item{dat}{observed data}

\item{lambda_min}{minimum lamba value}

\item{lambda_max}{maximum lamba value}

\item{max_iters}{maximum number of tuning parameters attempted}
}
\value{
Returns a list with elements:

\code{path_stats} a dataframe with summary statistics (number of changepoints, tuning parameters, cost)

\code{path_fits} a list containing the fitted model for each tuning parameter

\code{approximate_path} a boolean indicating whether an early stopping criterion condition occurred
}
\description{
Estimate changepoints using
automatic tuning parameter selection within
a range of values [lambda_min, lambda_max]
}
\details{
Uses the CROPS algorithm (https://www.tandfonline.com/doi/abs/10.1080/10618600.2015.1116445) to efficiently 
choose values of the tuning parameter between [lambda_min, lambda_max].
}
\examples{

set.seed(1)
mu <- rep(c(1, 3), each = 100)
dat <- mu + rnorm(length(mu))

fits <- estimate_paths(dat, lambda_min = 2, lambda_max = 3)
 
}
