% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{eval_cost_at.ChangepointInference_L0_changepoints_pvals}
\alias{eval_cost_at.ChangepointInference_L0_changepoints_pvals}
\title{Evaluate Cost(phi) for any phi}
\usage{
eval_cost_at.ChangepointInference_L0_changepoints_pvals(x, thj, phi)
}
\arguments{
\item{x}{ChangepointInference_L0_changepoints_pvals}

\item{thj}{changepoint}

\item{phi}{evaluate cost at pertubation phi}
}
\description{
Evaluate Cost(phi) for any phi
}
