% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{eval_cost_at.ChangepointInference_L0_estimated_changes}
\alias{eval_cost_at.ChangepointInference_L0_estimated_changes}
\title{Evaluate Cost_s*(mu) for any s and mu}
\usage{
eval_cost_at.ChangepointInference_L0_estimated_changes(x, s = 1,
  mu = 0)
}
\arguments{
\item{x}{ChangepointInference_L0_estimated_changes}

\item{s}{s in Cost_s*(mu), s = 1,..., length(dat)}

\item{mu}{mean cost to evaluate cost at}
}
\description{
Evaluate Cost_s*(mu) for any s and mu
}
