% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-noise-scale.R
\name{gen_diff_block_sd}
\alias{gen_diff_block_sd}
\title{Standard deviation estimator based on differences of partial sums of the data}
\usage{
gen_diff_block_sd(xx, ww = sqrt(length(xx)/2), degree = 0)
}
\arguments{
\item{xx}{data}

\item{ww}{TO DO!}

\item{degree}{degree of the underlying}
}
\description{
TO DO!
}
