% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-noise-scale.R
\name{gen_diff_catoni_tavc}
\alias{gen_diff_catoni_tavc}
\title{Robust estimation of TAVC with unknown piecewise polynomial signal}
\usage{
gen_diff_catoni_tavc(xx, ww, degree, tacv_max_scale = NULL, b_max = NULL)
}
\arguments{
\item{ww, }{scale at which TVAC will be calculated}

\item{degree}{polynomial degree of the underlying signal}

\item{tacv_max_scale, }{maximum scale at which to calculate TVAC}

\item{b_max}{number of starting points to use for the global TAVC estimator; helps with robustness.}

\item{xx_cumus, }{cumulative sum of the data}

\item{scaling}{numeric calling constant in local tests pre-computed by \code{diffInf}}
}
\description{
A short description...
}
\references{
McGonigle, Euan T., and Haeran Cho. "Robust multiscale estimation of time-average variance for time series segmentation." Computational Statistics & Data Analysis 179 (2023): 107648.

https://github.com/EuanMcGonigle/TAVC.seg
}
