% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-noise-scale.R
\name{gen_diff_mad}
\alias{gen_diff_mad}
\title{Median Absolute Deviation estimator based on differenced data}
\usage{
gen_diff_mad(xx, degree)
}
\arguments{
\item{xx}{data}

\item{degree}{degree of the underlying signal}
}
\description{
TO DO!
}
