% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-noise-scale.R
\name{gen_diff_sd}
\alias{gen_diff_sd}
\title{Standard deviation estimator based on differenced data}
\usage{
gen_diff_sd(xx, degree)
}
\arguments{
\item{xx}{data}

\item{degree}{degree of the underlying}
}
\description{
TO DO!
}
