% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-threshold.R
\name{H_nongauss}
\alias{H_nongauss}
\title{Get the numeric constant H under non-Gaussian and / or dependent noise}
\usage{
H_nongauss(degree, aa)
}
\arguments{
\item{degree}{polynomial degree of the underlying signal}

\item{aa}{controls grid density (a in the paper)}
}
\description{
Calculates the constant H_4 in the paper
}
