% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-threshold.R
\name{H_num_est}
\alias{H_num_est}
\title{H-num-est}
\usage{
H_num_est(degree, aa, LL)
}
\arguments{
\item{degree}{polynomial degree of the underlying signal}

\item{aa}{controls grid density (a in the paper)}

\item{LL}{control smallest scale (L in the paper)}
}
\description{
Numeric estimate of the constant H_2 in the paper
}
