% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-threshold.R
\name{H_num_est_gauss}
\alias{H_num_est_gauss}
\title{Estimate the numeric constant H under Gaussian noise}
\usage{
H_num_est_gauss(degree, aa)
}
\arguments{
\item{degree}{polynomial degree of the underlying signal}

\item{aa}{controls grid density (a in the paper)}
}
\description{
Numeric estimate of the constant H_2 in the paper
}
