% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-inf-bin-seg.R
\name{loc_diff}
\alias{loc_diff}
\title{Local differences}
\usage{
loc_diff(yy_cumsum, ll, ww, degree)
}
\arguments{
\item{yy_cumsum}{cumulative sum of data}

\item{ll}{start of the segment}

\item{ww}{number of data points in the segment}

\item{degree}{polynomial degree of the underlying signal}
}
\description{
Local differences
}
