% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrsd-block-diff.R
\name{lrsd_block_diff}
\alias{lrsd_block_diff}
\title{(Long Run) Standard Deviation Estimator Based on \eqn{(p+1)}-th Differences of Local Sums of the Data}
\usage{
lrsd_block_diff(
  yy,
  ww = length(yy)^{
     (1/3)
 },
  degree = 0
)
}
\arguments{
\item{yy}{A numeric vector containing the data.}

\item{ww}{The scale at which local sums of the data will be calculated.}

\item{degree}{The degree of polynomial parametrization of the mean of the data.}
}
\description{
Estimates the (long run) standard deviation based on \eqn{(p+1)}-th differences of non-overlapping local sums of the data.
}
\examples{
degree <- 3 

nn <- 500 

yy <- arima.sim(model = list(ar = 0.5), n = nn) + (1:nn)^{degree}

sd_diff(yy, degree)
}
