% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-threshold.R
\name{p_inf_sqrd}
\alias{p_inf_sqrd}
\title{P-inf}
\usage{
p_inf_sqrd(xx)
}
\arguments{
\item{xx}{a positive number}
}
\description{
Numerically approximates the square of function p_\inf on page 32 of Kabluchko (2018)
}
\references{
Kabluchko, Zakhar. "Extreme-value analysis of standardized Gaussian increments." arXiv preprint arXiv:0706.1849 (2018).
}
