% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{plot.ChangepointInference_BS_changepoints_pvals}
\alias{plot.ChangepointInference_BS_changepoints_pvals}
\title{Plot S as a function of \eqn{\phi}}
\usage{
\method{plot}{ChangepointInference_BS_changepoints_pvals}(x, thj,
  PLOT_MIN = -10, PLOT_MAX = 10, ...)
}
\arguments{
\item{x}{ChangepointInference_BS_changepoints_pvals}

\item{thj}{changepoint}

\item{PLOT_MIN}{minimum phi}

\item{PLOT_MAX}{maximum phi}
}
\description{
Plot S as a function of \eqn{\phi}
}
