% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{plot.ChangepointInference_L0_changepoints_pvals}
\alias{plot.ChangepointInference_L0_changepoints_pvals}
\title{Plot optimal cost and/or S as a function of \eqn{\phi}}
\usage{
\method{plot}{ChangepointInference_L0_changepoints_pvals}(x, thj,
  plot_cost = TRUE, PLOT_MIN = -10, PLOT_MAX = 10, ni = 1000, ...)
}
\arguments{
\item{x}{ChangepointInference_L0_changepoints_pvals}

\item{thj}{changepoint}

\item{plot_cost}{plot the optimal cost (TRUE), or plot S (FALSE)}

\item{PLOT_MIN}{minimum phi}

\item{PLOT_MAX}{maximum phi}

\item{ni}{number of values to calculate the optimal cost at in each segment}
}
\description{
Plot optimal cost and/or S as a function of \eqn{\phi}
}
