% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{plot.ChangepointInference_L0_estimated_changes}
\alias{plot.ChangepointInference_L0_estimated_changes}
\title{Plot optimal cost Cost_s*(mu) for any s}
\usage{
\method{plot}{ChangepointInference_L0_estimated_changes}(x, s = 1,
  PLOT_MIN = -10, PLOT_MAX = 10, ni = 1000, ...)
}
\arguments{
\item{x}{ChangepointInference_L0_estimated_changes}

\item{s}{s in Cost_s*(mu), s = 1,..., length(dat)}

\item{PLOT_MIN}{minimum phi to calculate cost}

\item{PLOT_MAX}{maximum phi to calculate cost}

\item{ni}{number of values to calculate the optimal cost at in each segment}
}
\description{
Plot optimal cost Cost_s*(mu) for any s
}
