% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cptInference}
\alias{plot.cptInference}
\title{Plot a 'cptInference' object}
\usage{
\method{plot}{cptInference}(obj, cpt_loc_est = FALSE, ...)
}
\arguments{
\item{obj}{An object of class 'cptInference', returned by \code{diffInf}.}
}
\description{
Plots the data and as well as intervals of significance returned by the algorithm
}
