% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{print.ChangepointInference_L0_changepoints_pvals}
\alias{print.ChangepointInference_L0_changepoints_pvals}
\title{Summary of estimated changepoints and p-values for L0 segmentation}
\usage{
\method{print}{ChangepointInference_L0_changepoints_pvals}(x, ...)
}
\arguments{
\item{x}{ChangepointInference_L0_changepoints_pvals object}
}
\description{
Summary of estimated changepoints and p-values for L0 segmentation
}
