% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{print.ChangepointInference_L0_estimated_changes}
\alias{print.ChangepointInference_L0_estimated_changes}
\title{Summary of estimated changepoints for L0 segmentation}
\usage{
\method{print}{ChangepointInference_L0_estimated_changes}(x, ...)
}
\arguments{
\item{x}{ChangepointInference_L0_estimated_changes object}
}
\description{
Summary of estimated changepoints for L0 segmentation
}
