% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-diff.R
\name{sd_diff}
\alias{sd_diff}
\title{Standard Deviation Estimator Based On \eqn{(p+1)}-th Differences}
\usage{
sd_diff(yy, degree)
}
\arguments{
\item{yy}{A numeric vector containing the data.}

\item{degree}{The degree of polynomial parametrization of the mean of the data.}
}
\description{
Estimates the standard deviation of the data using the mean on the squared \eqn{(p+1)}-th difference of the data sequence. Suitable for estimating the scale of the noise when the data sequence consists of a piecewise polynomial function contaminated with independently distributed noise having bounded fourth moment.
}
\examples{
degree <- 3 

nn <- 500 

yy <- rt(nn, df = 5) * sqrt(3/5) + (1:nn)^{degree}

sd_diff(yy, degree)
}
