% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt}
\alias{cpt}
\title{Estimate Change Point Locations}
\usage{
cpt(obj, cpt_loc = "RSS")
}
\arguments{
\item{obj}{An object of class \code{cptInference} returned by \code{\link{diffInf}}.}

\item{cpt_loc}{Determines how the change point locations will be estimated.
Should be one of "midpoint" or "RSS".
If \code{cpt_loc == "midpoint"} the change point locations are simply taken to be the midpoints of each interval.
If \code{cpt_loc == "RSS"} the change point locations are chosen to be the split points within each interval associated with the piecewsie polynomial fit providing the lowest sum of squared residuals.}
}
\value{
A vector of estimated change point locations.
}
\description{
Estimates the most likely change point location within each of the intervals obtained by calling \code{\link{diffInf}}.
}
\examples{
# Piecewise linear mean with i.i.d. Gaussian noise
set.seed(42)
 
waves_signal <- c((1:150) * (2**-3), (150:1) * (2**-3), (1:150) * (2**-3), (150:1) * (2**-3))
 
yy <- waves_signal + rnorm(length(waves_signal), sd = 5)
 
 
#' Recover intervals of significance
 
 diffInf_obj <- diffInf(yy, degree = 1)
 
 diffInf_obj
 
#' plot the intervals of significance
 
 diffInf_obj |> plot(type = "l", col = "grey")
 
 waves_signal |> lines(lty = 2, lwd = 2)
 
 
#' recover likely change point locations
 
cpt(diffInf_obj, "RSS")
 
cpt(diffInf_obj, "midpoint")
 
abline(v = cpt(diffInf_obj, "RSS"), col = "red", lty = 2)
 
abline(v = cpt(diffInf_obj, "midpoint"), col = "blue", lty = 3)
}
