% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-inf-bin-seg.R
\name{diff_bin_seg}
\alias{diff_bin_seg}
\title{diffInf binary segmentation}
\usage{
diff_bin_seg(xx_cumsum, ss, ee, degree, aa, min_scale, thresh, scaling)
}
\arguments{
\item{xx_cumsum}{cumulative sum of data}

\item{ss}{start of search area}

\item{ee}{end of search area}

\item{degree}{polynomial degree of the underlying signal}

\item{aa}{controls grid density (a in the paper)}

\item{min_scale}{minimum width of intervals tested for a change}

\item{thresh}{threshold to control FWE computed by \code{diffInf}}

\item{scaling}{numeric calling constant in local tests pre-computed by \code{diffInf}}
}
\description{
Recursively search for narrowest interval containing an change point over a particular logarithmic grid using binary segmentation algorithm; not not be called directly by the user.
}
