% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffInf.R
\name{diffInf}
\alias{diffInf}
\title{Change point inference via differencing}
\usage{
diffInf(xx, degree, alpha, sigma = NULL, aa = 2, LL = 0, HH = NULL)
}
\arguments{
\item{xx}{signal to be segmented}

\item{degree}{polynomial degree of the underlying signal}

\item{alpha}{desired coverage}

\item{sigma}{noise level if know, if null sigma will be estimated using}

\item{aa}{controls grid density (a in the paper)}

\item{LL}{control smallest scale (L in the paper)}

\item{HH}{supply numeric constant H if pre-computed}
}
\description{
Main routine
}
