% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised-tacv-est.R
\name{generalised_tavc_est}
\alias{generalised_tavc_est}
\title{Robust estimation of TAVC with unknown piecewise polynomial signal}
\usage{
generalised_tavc_est(
  xx_cumsum,
  ww,
  degree,
  scaling,
  tacv_max_scale = NULL,
  b_max = NULL
)
}
\arguments{
\item{ww, }{scale at which TVAC will be calculated}

\item{degree}{polynomial degree of the underlying signal}

\item{scaling}{numeric calling constant in local tests pre-computed by \code{diffInf}}

\item{tacv_max_scale, }{maximum scale at which to calculate TVAC}

\item{b_max}{number of starting points to use for the global TAVC estimator; helps with robustness.}

\item{xx_cumus, }{cumulative sum of the data}
}
\description{
A short description...
}
\references{
McGonigle, Euan T., and Haeran Cho. "Robust multiscale estimation of time-average variance for time series segmentation." Computational Statistics & Data Analysis 179 (2023): 107648.

https://github.com/EuanMcGonigle/TAVC.seg
}
