% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-threshold.R
\name{get_thresh}
\alias{get_thresh}
\title{Get threshold}
\usage{
get_thresh(nn, WW, alpha, degree, aa, HH, gauss_indep_noise)
}
\arguments{
\item{nn}{length of the data sequence}

\item{WW}{minimum segment length}

\item{alpha}{desired coverage}

\item{degree}{polynomial degree of the underlying signal}

\item{aa}{controls grid density (a in the paper)}

\item{HH}{numeric constant}

\item{gauss_indep_noise}{true if noise is Gaussian and independent}
}
\description{
Calculate threshold for FWE control
}
