% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-threshold.R
\name{get_thresh}
\alias{get_thresh}
\title{Get threshold}
\usage{
get_thresh(nn, alpha, degree, aa, LL, HH)
}
\arguments{
\item{nn}{length of the data sequence}

\item{degree}{polynomial degree of the underlying signal}

\item{aa}{controls grid density (a in the paper)}

\item{LL}{control smallest scale (L in the paper)}

\item{HH}{numeric constant}
}
\description{
Calculate threshold for FWE controll
}
