% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cptInference}
\alias{predict.cptInference}
\title{Predict a 'cptInference' object}
\usage{
\method{predict}{cptInference}(obj, cpt_loc = "RSS")
}
\arguments{
\item{obj}{An object of class 'cptInference', returned by \code{diffInf}.}
}
\description{
123
}
